// src/types/socket/VehicleFms.ts
interface VehicleFms {
  // eslint-disable-next-line @typescript-eslint/naming-convention
  userVehicleID: number;
  userVehicleName: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  userVehicleFMS: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  userMissionID: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  userDepartmentID: number;
  countPersonal: number;
  vehicleShortName: string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  vehicleID: number;
}

export type { VehicleFms };